Move gammapy.cube.models to gammapy.modeling.models.cube
