+++ docs/modeling/models.rst
@@ -1,1 +1,29 @@
+.. _modeling_models:
+
+************************
+modeling.models - Models
+************************
+
+.. currentmodule:: gammapy.modeling.models
+
+Introduction
+============
+
 tbd
+
+Getting Started
+===============
+
+tbd
+
+Using `gammapy.modeling.models`
+===============================
+
+tbd
+
+Reference/API
+=============
+
+.. automodapi:: gammapy.modeling.models
+    :no-inheritance-diagram:
+    :include-all-objects:
