+++ gammapy/cube/fit.py
@@ -10,7 +10,7 @@ from ..utils.fitting import Parameters, Dataset
 from ..stats import cash, cstat, cash_sum_cython, cstat_sum_cython
 from ..maps import Map
 from ..irf import EnergyDispersion
-from .models import SkyModel, SkyModels, BackgroundModel
+from gammapy.modeling.models import SkyModel, SkyModels, BackgroundModel
 from .psf_kernel import PSFKernel
