+++ gammapy/modeling/models/__init__.py
@@ -1,4 +1,12 @@
-"""Built-in image in Gammapy"""
+"""Models in Gammapy.
+
+TBD, later: how to make model names uniform?
+- "Spectral" for ``gammapy.modeling.models.spectrum``
+- "Sky" or "Image" ? for ``gammapy.modeling.models.image``
+- "Sky" or "Cube" ? for ``gammapy.modeling.models.cube``
+- "Time" ? for ``gammapy.modeling.models.time``
+"""
 from .spectrum import *
 from .image import *
+from .cube import *
 from .time import *
